+++ src/components/layout/NavPage.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import Genres from "../pages/Genre";
-import LibraryPatron from "../pages/LibraryPatron";
-import Book from "../pages/Book";
+import Genres from "../pages/Admin/Genre";
+import LibraryPatron from "../pages/Admin/LibraryPatron";
+import Book from "../pages/Admin/Book";
 
 const NavPage = () => {
   return (
@@ -10,6 +10,7 @@ const NavPage = () => {
       <section className="bg-white m-4 p-5 rounded font-bold text-2xl">
         <div>
           <Routes>
+            <Route path="/" element={<Book />} />
             <Route path="/genres" element={<Genres />} />
             <Route path="/patrons" element={<LibraryPatron />} />
             <Route path="/books" element={<Book />} />
